<! DOCTYPE html > <html> <head> <title>Assignment </title> </head > <body> <table border="1"> <caption><h1>Practice Table</h1></caption> <tr> <td> <table border="1"> <tr> <td><img src="https://www.gardendesign.com/pictures/images/675x529Max/site_3/helianthus-yellow-flower-pixabay_11863.jpg" height="100%" width="100%"</td> <td><img src="https://hips.hearstapps.com/hmg-prod.s3.amazonaws.com/images/close-up-of-tulips-blooming-in-field-royalty-free-image-1584131603.jpg" height="40%" width="30%"</td> </tr> <tr> <td><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ7DSG6qgN7-BNccysyV5QZxMXsMJngt_8gqOyj4yCo1YEy068ilZm5xVFO1w6TYzJ3cUI&usqp=CAU" height="100%" width="100%"</td> <td><img src="https://imgcdn.floweraura.com/FA-12-CCC.jpg" height="40%" width="30%" </td> </tr> </table> </td> <td> <table border="1"> <tr> <td> <div style="background-color:yellow"> <a href="https://www.goodhousekeeping.com/holidays/mothers-day/g5187/mothers-day-flowers/">Cakes</a> </div> </td> <td> <div style="background-color:green;"> <a href="https://www.bigbasket.com/pc/snacks-branded-foods/chocolates-candies/">Chocolates</a> </div> </td> </tr> <tr> <td> <div style="background-color:pink"> <a href="https://kcroasters.com/?gclid=Cj0KCQiAtJeNBhCVARIsANJUJ2FeqyPNlnfX7Uko3x1aOD-EBQKAZVTUXEQOhp2SZwNcC-LmSUV1rIcaAgXUEALw_wcB">Coffee</a> </div> </td> <td> <div style="background-color:orange;"> <a href="https://www.hungryhowies.com/?gclid=Cj0KCQiAtJeNBhCVARIsANJUJ2HtETR7XD84Tb-_vS78w-le6TWGrlqJMS_tEB3XCRAUDaQrsLc3suIaAu0GEALw_wcB&gclsrc=aw.ds">Pizza</a> </div> </td> </tr> </table></td> </tr> <tr> <td> <table border="1"> <p><div style="background-color:black;color:white;"> Sunset, also known as sundown, is the daily disappearance of the Sun below the horizon due to Earth's rotation. As viewed from everywhere on Earth (except the North and South poles), the equinox Sun sets due west at the moment of both the Spring and Autumn equinox. As viewed from the Northern Hemisphere, the sun sets to the northwest (or not at all) in the Northern hemisphere's spring and summer, and to the southwest in the autumn and winter; these seasons are reversed for the Southern Hemisphere.
The time of sunset is defined in astronomy as the moment when the upper limb of the Sun disappears below the horizon.[1] Near the horizon, atmospheric refraction causes sunlight rays to be distorted to such an extent that geometrically the solar disk is already about one diameter below the horizon when a sunset is observed.

Sunset is distinct from twilight, which is divided into three stages. The first one is civil twilight, which begins once the Sun has disappeared below the horizon, and continues until it descends to 6 degrees below the horizon. The second phase is nautical twilight, between 6 and 12 degrees below the horizon. As for the third one, it is astronomical twilight, which is the period when the Sun is between 12 and 18 degrees below the horizon.[2] Dusk is at the very end of astronomical twilight, and is the darkest moment of twilight just before night.[3] Finally, night occurs when the Sun reaches 18 degrees below the horizon and no longer illuminates the sky.
</div>
</P>
</table>
</td>
<td>
<table border="1">
<p>
<div style="background-image:url('https://royalpepper.in/blog/wp-content/uploads/2021/08/f1c1c93f3f0266e714d071a5f9ba273802ba7c64.jpg');"
height="50%"
width="50%">
India, officially the Republic of India (Hindi: Bhārat Gaṇarājya),[23] is a country in South Asia.
It is the seventh-largest country by area, the second-most populous country, and the most populous democracy in the world.
Bounded by the Indian Ocean on the south, the Arabian Sea on the southwest, and the Bay of Bengal on the southeast, it shares land borders with Pakistan to
 the west;[f] China, Nepal, and Bhutan to the north; and Bangladesh and Myanmar to the east.
In the Indian Ocean, India is in the vicinity of Sri Lanka and the Maldives; its Andaman and Nicobar Islands share a maritime border with Thailand, Myanmar and Indonesia.
Modern humans arrived on the Indian subcontinent from Africa no later than 55,000 years ago.[24] Their long occupation, initially in varying forms of isolation as hunter-gatherers, has made the region highly diverse, second only to Africa in human genetic diversity.[25] Settled life emerged on the subcontinent in the western margins of the Indus river basin 9,000 years ago, evolving gradually into the Indus Valley Civilisation of the third millennium BCE.[26] By 1200 BCE, an archaic form of Sanskrit, an Indo-European language, had diffused into India from the northwest,[27] unfolding as the language of the Rigveda, and recording the dawning of Hinduism in India.[28] The Dravidian languages of India were supplanted in the northern and western regions.[29] By 400 BCE, stratification and exclusion by caste had emerged within Hinduism,[30] and Buddhism and Jainism had arisen, proclaiming social orders unlinked to heredity.[31] Early political consolidations gave rise to the loose-knit Maurya and Gupta Empires based in the Ganges Basin.[32] Their collective era was suffused with wide-ranging creativity,[33] but also marked by the declining status of women,[34] and the incorporation of untouchability into an organised system of belief.[g][35] In South India, the Middle kingdoms exported Dravidian-languages scripts and religious cultures to the kingdoms of Southeast Asia.
</div>
</p>
</table>
</td>
</table>

</body>
</html>